third_party/blink/web_tests/external/wpt/payment-handler/can-make-payment-event.https-expected.txt

This is a testharness.js-based test.
[FAIL] If a payment handler is not installed, but can be installed just-in-time, then the payment method is supported.
  assert_equals: If it throws, then it must be NotAllowedError expected "NotAllowedError" but got "UnknownError"
[FAIL] If a payment handler is not installed and cannot be installed just-in-time, then the payment method is not supported.
  assert_equals: If it throws, then it must be NotAllowedError expected "NotAllowedError" but got "UnknownError"
[FAIL] If CanMakePaymentEvent.respondWith(false) is called, then canMakePayment() still returns true and the app can still be invoked.
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'show' on 'PaymentRequest': Already called show() once"
[FAIL] If CanMakePaymentEvent.respondWith(Promise.resolve(false)) is called, then canMakePayment() still returns true and the app can still be invoked.
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'show' on 'PaymentRequest': Already called show() once"
[FAIL] If CanMakePaymentEvent.respondWith(true) is called, then canMakePayment() returns true and the app can be invoked.
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'show' on 'PaymentRequest': Already called show() once"
[FAIL] If CanMakePaymentEvent.respondWith(Promise.resolve(true)) is called, then canMakePayment() returns true and the app can be invoked.
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'show' on 'PaymentRequest': Already called show() once"
[FAIL] If CanMakePaymentEvent.respondWith(Promise.reject(error)) is called, then canMakePayment() still returns true and the app can still be invoked.
  promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'show' on 'PaymentRequest': Already called show() once"
Harness: the test ran to completion.
